requirements.txt
flask
flask-cors
joblib
numpy
pandas
easyocr
Pillow
opencv-python-headless
torch